refactored to move to a new package raccoon_client
