+++ clients/python/raccoon_client/serde/enum.py
@@ -1,7 +1,7 @@
 from enum import Enum
 
-from serde.json_serde import JsonSerde
-from serde.protobuf_serde import ProtobufSerde
+from raccoon_client.serde.json_serde import JsonSerde
+from raccoon_client.serde.protobuf_serde import ProtobufSerde
 
 
 class Serialiser(Enum):
